Rename multirust_dist::install multirust_dist::prefix
It only holds InstallPrefix now.
