+++ src/multirust-dist/tests/dist.rs
@@ -14,7 +14,7 @@ extern crate hyper;
 
 use multirust_mock::dist::*;
 use multirust_mock::{MockCommand, MockInstallerBuilder};
-use multirust_dist::install::InstallPrefix;
+use multirust_dist::prefix::InstallPrefix;
 use multirust_dist::{Error, NotifyHandler};
 use multirust_dist::dist::ToolchainDesc;
 use multirust_dist::download::DownloadCfg;
